+++ example/console_example.dart
@@ -18,7 +18,7 @@ Future<void> main() async {
     const Str('mime', 'image/png'),
   });
 
-  final tracer = context.trace('Uploading!', Level.info);
+  final tracer = context.trace('Uploading!', level: Level.info);
 
   // Emulate uploading, wait for 1 sec.
   await Future<void>.delayed(const Duration(seconds: 1));