+++ lib/src/interface.dart
@@ -14,7 +14,7 @@ abstract class Interface {
 
   /// Starts tracing and emits a record with [message] and [level]
   /// severity level; to stop tracing call [Tracer.stop] on the returned tracer.
-  Tracer trace(String message, [Level level = Level.debug]);
+  Tracer trace(String message, {Level level = Level.debug});
 
   /// Creates and returns a new logging context with bound collection of
   /// [fields] added to existing one.