+++ lib/src/tracer.dart
@@ -1,4 +1,6 @@
+import 'package:jetlog/src/field.dart' show Field;
 import 'package:jetlog/src/interface.dart';
+import 'package:jetlog/src/level.dart' show Level;
 
 /// Tracer is used to measure time between [Interface.trace] and [Tracer.stop]
 /// calls.
@@ -8,5 +10,5 @@ abstract class Tracer {
   ///
   /// If [stop] is called more than once a [TracerStoppedError]
   /// will be raised.
-  void stop(String message);
+  void stop(String message, {Level? level, Iterable<Field>? fields});
 }